requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
cssselect==1.2.0
cssutils==2.11.1
idna==3.10
lxml==5.3.0
more-itertools==10.5.0
premailer==3.10.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
yagmail==0.15.293
